requirements.txt
openai==0.27.0
pdfplumber
python-dotenv
streamlit